requirements.txt
pandas==1.4.2
pandas-ta==0.3.14b0
pymongo==4.1.1
python-binance==1.0.16
python-dotenv==0.20.0